rename lib/pages.jsx to lib/routes.jsx.
